packages/eva/captioner/requirements.txt
fastapi>=0.115,<1.0
uvicorn[standard]>=0.32,<1.0
dynaconf>=3.2,<4.0
Pillow>=11.0,<12.0
torch>=2.2,<2.4
torchvision>=0.17,<0.19
transformers>=4.46,<5.0
